+++ CHANGES.rst
@@ -5,6 +5,13 @@ Changelog
 ------------------
 
 - #2573 Move Interpretation Templates to Senaite setup folder
+- #2572 Move Sample Containers to Senaite setup folder
+- #2571 Move InstrumentLocations to Senaite setup folder
+- #2578 Fix setup DX folders are not indexed in `uid_catalog`
+- #2576 Fix analyses and services are re-indexed after modifying category
+- #2577 Fix content types typo
+- #2567 Migrate AnalysisCategories to Dexterity
+- #2570 Use reference widget API in sample add form
 - #2553 Migrate BatchLabels to Dexterity
 - #2566 Migrate SamplePoint Latitude and Longitude fields to Location
 - #2552 Migrate SamplingDeviations to Dexterity